add test/ and testasmcomp/ from ocaml github repo at 1.07!
The ocaml-1.07.tgz I based ocaml-light on actually was the official
distrib of OCaml but didn't contain all the code in the
OCaml (CVS at the time) repo because tests were not useful for people
installing OCaml. At some point though OCaml got available on github
and we got the full history and could see the additional dirs!