fix configure to work on my arch linux
Not sure why but I was also getting new warnings and errors
in configure preventing to compile correctly. Actually I think
it never worked but I had an existing config/s.h and config/m.h
that was working.

Anyway now configure works and so are the tests and everything