ugly hack to compile otherlibs/{str,unix} with -m32 too
the otherlibs/unix/libunix.a was compiled with BYTECC
but it can't then be linked with objects produced by ocamlopt
because of the missing -m32
=> recompile libunix also with -32 and add the opt suffix
(libunixopt.a) to differentiate.
That means the caller needs to use -cclib -lunixopt
but it's ok for now.

test plan:
can compile now xix with ocamlight opt!